loja-cupcake/Projeto/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.1.1
python-dotenv==1.0.0
mysql-connector-python==8.0.33
pytest==7.4.3
selenium==4.15.2
requests==2.31.0
